+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.131] - 2025-10-02
+### Added
+* overloaded function call to `EditorUtils`
+* comments to `EditorUtils`
+* `IListExtensions.GetRandomIndex`
+### Changed
+* call to `SerializedObject.Update` to `SerializedObject.UpdateIfRequiredOrScript` within `EditorUtils`
+
 ## [0.0.130] - 2025-09-17
 ### Fixed
 * accidental `DOTween` reference include